Handle cases where an SshChannel has been free'd from C
If the remote end disconnects it seems that libssh will disconnect/free the
`ssh_session`'s channels. This can lead to an `SshChannel` carrying around a
pointer to an address that's already been free'd, which leads to segfaults. Now
we always check if the session has been disconnected when checking if a
`SshChannel` is open and when closing it.