Refactor wait(::Session) so that the polling happens asynchronously
Previously each call to `wait(::Session)` would poll the sessions file
descriptor individually. This is kinda not great because one caller may be woken
up for some data that's intended for another caller. In practice this hasn't
caused problems so far because most of LibSSH's usage has been in the tests
where we do a single operation at a time, but the new architecture should be
safer.

Now there's a separate waiter task that lives for as long as the session that
can be woken up by callers wanting to wait() on the session, who will in turn be
woken up whenever the waiter tasks polls the session file descriptor
successfully. i.e. it turns the waiting from a 1-to-1 thing to a 1-to-many
thing.